Final submission
With this project, we explore some topics around Java Collections. It's intention is to create "Flashcards" to help create and practice certain information pairs: could be for learning a language, or for memorising countries' capitals etc. It has the ability to import an existing text file that contains information that can be parsed to the programme, as well as exporting to a text file after creating it. It was a good practice for things like loops, ArrayList, HashMap, Scanner, loops, and FileWriter